EnvironmentalCompany
This work was performed during participation in the competition from the iTeen Academy - "Grodno Team Competition". Where the work I created got the first place.
The first task was to make up the page according to the layout with the resources given to us. After completing the first task, the second task was to make an adaptive version for mobile devices.